refactor(agentboot): fold session/ micro-package into the root package
The session package was a single 19-line interface (LifecycleStore) plus a
deprecated alias, with exactly two importers. The interface now lives in
agentboot/lifecycle.go next to ExecutionOptions.Store which carries it; the
deprecated session.Store alias is gone with the package. The compile-time
assertion in remote/control/adapter/session_bridge.go now checks against
agentboot.LifecycleStore.
